+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - configuring time zones for users and companies
+- average time calculations for night working time
 
 ## [1.3.2] - 2023-06-08
 ### Added